% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolSineNL.R
\name{convolSineNL}
\alias{convolSineNL}
\title{Convolution of the sine function with the gamma function}
\usage{
convolSineNL(
  AP = NULL,
  phiP = NULL,
  kP = NULL,
  estAlpha = NULL,
  estBeta = NULL,
  simulatedDate = NULL,
  nWarmupYears = NULL,
  printAll = FALSE
)
}
\arguments{
\item{AP}{Amplitude of the sine wave describing isotope in precipitation: \cr
\eqn{c_P = A_P \cdot sin(2 \pi t - \varphi_P) + k_P}}

\item{phiP}{Phase shift of the sine wave describing isotope in precipitation
(please see the above equation).}

\item{kP}{The constant factor of sine wave describing isotope in precipitation
(please see the above equation).}

\item{estAlpha}{The 'user' estimated alpha (shape) of the gamma distribution
function used for the convolution approach}

\item{estBeta}{The estimated beta (scale) of the gamma distribution function
use for the convolution approach}

\item{simulatedDate}{The date vector (must be in format "YYYY-MM-DD") and as
DATE of the time tracer concentration in streamflow should be estimated}

\item{nWarmupYears}{Number of warm-up years before the starting date of the
simulatedDate. This should be a few years (e.g.,5 years) to remove the effect
of initial condition on the estimated results.}

\item{printAll}{A logical variable, if TRUE then the results (simulated isotope
concentration in streamflow) is printed for the entire simulation period,
including the warm up period. If FALSE then the results is only for the
simulatedDate.}
}
\value{
A data frame containing the date and simulated tracer concentrations.
}
\description{
Convolution of the sine-wave function describing isotope in precipitation with
the gamma (transit time) distribution function.
}
\details{
This function perform the following convolution (please see Eq. 1 from Kirchner (2016)):

\deqn{c_S(t) = \int_{0}^{\infty} h(\tau) \cdot c_P(t - \tau) d\tau }{}

where \eqn{c_P(t)} in the following sine-wave form: \cr

 \deqn{c_P(t) = A_P \cdot sin(2 \pi t - \varphi_P) + k_P}{}

and the gamma distribution is:

\deqn{h(\tau) =  \frac{\tau^{\alpha - 1}}{\beta^{\alpha} \Gamma (\alpha)} e^ {-\tau/\beta}}{}
}
\examples{
# Get tracer data in streamflow of the Alp catchment (from the example dataset)
isotopeS_Alp <- subset(isotopeData, catchment == "Alp" & variable == "streamflow")

\dontrun{
simC <- convolSineNL(AP = 1.909, phiP = 2.016, kP = -11.05995, estAlpha = 0.802,
                     estBeta = 0.127, simulatedDate = isotopeS_Alp$date,
                     nWarmupYears = 5, printAll = FALSE)
}
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297
}
