% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlpha.R
\name{findAlpha}
\alias{findAlpha}
\title{Find alpha value using bisection approach}
\usage{
findAlpha(phiS = NULL, phiP = NULL, Fyw = NULL, eps = 1e-06, maxIter = 1000)
}
\arguments{
\item{phiS}{Phase shift of isotope concentration in streamflow (in radian). \cr
phiS MUST be within the range of [0, 2*pi] \cr
phiS MUST be smaller than phiP}

\item{phiP}{Phase shift of isotope concentration in precipitation (in radian).
phiP MUST be within the range of [0, 2*pi]}

\item{Fyw}{Youngwater fraction, equals to the amplitude ratio AS/AP where AS
and AP are the amplitude of the fitted sine wave to isotope concentrations in
streamflow and in rainfall, respectively.}

\item{eps}{The minimum search distance}

\item{maxIter}{Maximum number of iteration when using bisection approach}
}
\description{
Find alpha value using bisection approach
}
\examples{
findAlpha <- function(phiS = 2*pi, phiP = pi, Fyw = 0.3, eps = 1e-6,
maxIter = 1000)
}
