% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlphaBeta.R
\name{findAlphaBeta}
\alias{findAlphaBeta}
\title{Find findAlphaBeta using bisection approach}
\usage{
findAlphaBeta(phiS = NULL, phiP = NULL, Fyw = NULL, eps = 1e-06)
}
\arguments{
\item{phiS}{Phase shift of isotope concentration in streamflow (in radian). \cr
phiS MUST be within the range of [0, 2*pi] \cr
phiS MUST be smaller than phiP (will be automatically checked inside this function)}

\item{phiP}{Phase shift of isotope concentration in precipitation (in radian).
phiP MUST be within the range of [0, 2*pi]}

\item{Fyw}{Youngwater fraction, equals to the amplitude ratio AS/AP where AS
and AP are the amplitudes of the fitted sine wave to isotope concentrations in
streamflow and in precipitation, respectively.}

\item{eps}{The minimum search distance}

\item{maxIter}{Maximum number of iteration when using bisection approach}
}
\value{
A list object containing the shape (alpha), scale (beta) factors,
and also mean transit time
}
\description{
Find alpha (shape) and beta (scale) factor of the gamma
distriubtion using the bisection approach. Solve equations (10) and (11) in
Kirchner (2016).
}
\examples{
result <- findAlphaBeta(phiS = 2*pi,
                        phiP = pi,
                        Fyw = 0.3,
                        eps = 1e-6)
# alpha (shape) factor
result$alpha
# beta (scale) factor
result$beta
# Mean transit time (years)
result$meanTT
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
