% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlphaBetaMTT.R
\name{findAlphaBetaMTT}
\alias{findAlphaBetaMTT}
\title{Find alpha, beta, and mean transit time}
\usage{
findAlphaBetaMTT(phiS_phiP = NULL, AS = NULL, AP = NULL, eps = 1e-06)
}
\arguments{
\item{phiS_phiP}{Phase shift of isotope concentration in streamflow (in radian).
phiS_phiP MUST be > 0}

\item{AS}{The amplitude of the sine-wave function of isotope concentration in
precipitation. AS MUST be > 0}

\item{AP}{The amplitude of the sine-wave function of isotope concentration in
streamflow. AP MUST be > 0}

\item{eps}{The minimum search distance}
}
\value{
A table object containing the alpha, beta, and mean transit time.
}
\description{
Find alpha (shape) and beta (scale) factors of the gamma (transit time)
distribution using the bisection approach. Solve equations (10) and (11) in
Kirchner (2016).
}
\examples{
result <- findAlphaBetaMTT(phiS_phiP = pi,
                        AS = 1.0,
                        AP = 2.0,
                        eps = 1e-6)
result
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems –
Part 1: Seasonal tracer cycles quantify young water fractions, but not mean
transit times, in spatially heterogeneous catchments. \emph{Hydrol. Earth Syst. Sci.},
20, 279–297. \url{https://doi.org/10.5194/hess-20-279-2016}.
}
