% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPhi.R
\name{findPhi}
\alias{findPhi}
\title{Find \eqn{\varphi} within the range of \eqn{[0, 2\pi]}}
\usage{
findPhi(aP = NULL, bP = NULL)
}
\arguments{
\item{aP}{A scalar or vector of the shape factor of the gamma distribution}

\item{bP}{A scalar or vector of the scale factor of the gamma distribution
phiS MUST be smaller than phiP}
}
\value{
A single phiP value within the range of \eqn{[0, 2\pi]}
}
\description{
Find \eqn{\varphi} within the range of \eqn{[0, 2\pi]}
}
\details{
This function finds root of the following system of equations for \eqn{\varphi_P}
that within the range of \eqn{[0, 2\pi]}{} given values of aP and bP \cr
\deqn{a_P = - \sqrt{a_P^2 + b_P^2} \cdot sin(\varphi_P) = -A_p \cdot sin(\varphi_P)}{}
\deqn{b_P = \sqrt{a_P^2 + b_P^2} \cdot cos(\varphi_P) = A_p \cdot cos(\varphi_P))}{}
}
\examples{
Ap <- 2.5
phiP <- 0.1
findPhi(aP = -Ap*sin(phiP), bP = Ap*cos(phiP))
# This function should return phiP = 0.1
phiP <- 2*pi - 0.1
findPhi(aP = -Ap*sin(phiP), bP = Ap*cos(phiP))
# This function should return phiP = 2*pi - 0.1
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
