% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGamma.R
\name{fitGamma}
\alias{fitGamma}
\title{Fit gamma distribution to observed instream tracer concentratoin}
\usage{
fitGamma(
  AP = NULL,
  phiP = NULL,
  kP = NULL,
  alphaRange = NULL,
  betaRange = NULL,
  simulatedDate = NULL,
  fittedData = NULL,
  weight = NULL,
  nIter = 5000,
  nBestIter = 30,
  nCores = 1,
  nWarmupYears = NULL
)
}
\arguments{
\item{AP}{Amplitude of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
isotope concentration in precipitation.}
}
\value{
A data frame of the date and simulated instream isotope concentration
}
\description{
Fit gamma distribution to observed instream tracer concentratoin
}
\details{
This function convolute the input sine wave (cP = AP*sin(2*pi*t - phiP)
+ kP) with the given gamma distribution (user given the shape and scale factor)
and return the simulated isotope concentrations in streamflow
}
\examples{

#Get isotope data in streamflow of the Alp catchment (from the example dataset)
isotopeS_Alp <- subset(isotopeData, catchment == "Alp" & variable == "streamflow")

simC <- fitGamma(AP = 1.909, phiP = 2.016, kP = -11.05995, alphaRange = c(0.01,5),
                     betaRange = c(0.01,10), simulatedDate = isotopeS_Alp$date,
                     fittedData = isotopeS_Alp$delta_18O,
                     weight = isotopeS_Alp$water_flux_mm, nIter = 8,
                     nBestIter = 3, nCores = 1,
                     nWarmupYears = 5)
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
