% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSineNL.R
\name{fitSineNL}
\alias{fitSineNL}
\title{Fit a nonlinear sine wave function}
\usage{
fitSineNL(
  obsC = NULL,
  a = c(0, 10),
  phi = c(0, 2 * pi),
  k = c(-20, 0),
  t = NULL,
  nIter = 50000,
  nBestIter = 30,
  weights = rep(1, length(obsC))
)
}
\arguments{
\item{obsC}{A numeric vectors containing values of the dependent variables, this
could be measure isotope concentration in precipitation (or streamflow).}

\item{phi}{A numeric vector, representing the maximum and minimum search range
for the phase shift of isotope concentrations in precipitation (or
streamflow). The range must be within [0, 2*pi].}

\item{k}{A numeric vector, representing the maximum and minimum search range
for the phase shift of the constant value in the fitted function.}

\item{t}{A date (or character) vector in format format yyyy-mm-dd
(e.g., 2020-12-31)representing the date of observed isotope concentrations}

\item{A}{A numeric vector with positive numbers, representing the maximum and
minimum search range for the amplitude A of cycle of isotope concentrations
in precipitation (or streamflow).}
}
\value{
A list of object
}
\description{
Fit a nonlinear sine wave function
}
\details{
Fitting non linear sine function of the following form: \cr
y = a * sin(2 * pi * t - phi) + k
}
\examples{
fitSineNL(y = runif(10), t = runif(10))
}
\seealso{
toDecimal, randomLHS
}
