% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fyw.R
\name{Fyw}
\alias{Fyw}
\title{Find the youngwater fraction (Fyw) and other associtate parameters}
\usage{
Fyw(AP = NULL, phiP = NULL, AS = NULL, phiS = NULL)
}
\arguments{
\item{AP}{A numeric vectors (or a scalar), amplitude of sine wave of}
}
\value{
A list of object
}
\description{
Find the youngwater fraction (Fyw) and other associtate parameters
}
\details{
Fitting non linear sine function of the following form:
}
\examples{
Fyw(AP = runif(3), phiP = runif(3), AS = runif(4), phiS = runif(4))
}
\seealso{
findAlphaBeta
}
