% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively reweighted least squares (IRLS) with optional point weights}
\usage{
IRLS(
  Y,
  X,
  pweights = rep(1, length(Y)),
  iweights = rep(1, length(Y)),
  type = "Cauchy"
)
}
\arguments{
\item{Y}{A numeric vector representing a response variable}

\item{X}{A numeric vector or array representing one or more explanatory variables.
If X is a numeric vector, X must have the same length as Y.
If X is an array, the first dimension of X mas have the same length
as Y. Y and X must not be exactly collinear. This is not checked.}

\item{pweights}{An optional numeric vector of point weights (such as masses for
mass-weighted regressions) the default weight vector gives all pweights
the value of 1}

\item{iweights}{An optional numeric vector of initial values for the residual
weights for the IRLS, the default weight vector gives all initial
iweights's the value of 1}

\item{type}{An optional string constant indicating the weight function.
the value of type could be "bisquare", "Welsch", or "Cauchy". Default
weight function is Cauchy.}
}
\value{
Object of class "lm", generated by a call
        with the initial weighted updated by the selected weighting function
        fit$iweights
}
\description{
Iteratively reweighted least squares (IRLS) with optional point weights
}
\examples{
# Explanatory variables
X <- c(1:100) + runif(100)/10
# Response variable
Y <- X + runif(100)*50
# Point weights
pweights <- rep(1, length(X))
# Initial weights
iweights <- rep(1, length(X))
results <- IRLS(Y = Y, X = X, pweights = pweights,
                iweights = iweights, type = "Cauchy")
summary(results)
results$iweights
}
\references{
Kirchner, J. W.: Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016, 2016.

von Freyberg, J., Allen, S. T., Seeger, S., Weiler, M., and Kirchner, J. W.:
Sensitivity of young water fractions to hydro-climatic forcing and landscape
properties across 22 Swiss catchments, Hydrol. Earth Syst. Sci., 22,
3841–3861, https://doi.org/10.5194/hess-22-3841-2018, 2018.
}
