% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopeQ}
\alias{isotopeQ}
\title{Isotope (O18) in streamflow from the Erlenbach catchment (Switzerland)}
\description{
Isotope (O18) in streamflow from the Erlenbach catchment (Switzerland)
}
\references{
\url{https://www.envidat.ch/#/metadata/four-years-of-daily-stable-water-isotope-data}
}
\author{
Tam V. Nguyen \email{nguyen@ufz.de}
}
\keyword{Isotope}
\keyword{in}
\keyword{streamflow}
