% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolSineNL.R
\name{convolSineNL}
\alias{convolSineNL}
\title{Convolve the sine-wave function with the gamma distribution function}
\usage{
convolSineNL(
  AP = NULL,
  phiP = NULL,
  kP = NULL,
  estAlpha = NULL,
  estBeta = NULL,
  simulatedDate = NULL,
  nWarmupYears = NULL,
  printAll = FALSE
)
}
\arguments{
\item{AP}{Amplitude of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
tracer concentration in precipitation.}

\item{phiP}{Phase shift of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
tracerconcentration in precipitation.}

\item{kP}{The constant factor of sine wave (cP = AP*sin(2*pi*t - phiP) + kP)
of tracer concentration in precipitation.}

\item{estAlpha}{The 'user' estimated alpha (shape) of the gamma distribution
function used for the convolution approach}

\item{estBeta}{The estimated beta (scale) of the gamma distribution function
use for the convolution approach}

\item{simulatedDate}{The date vector (must be in format "YYYY-MM-DD") and as
DATE of the time tracer concentration in streamflow should be estimated}

\item{nWarmupYears}{Number of warm-up years before the starting date of the
simulatedDate. This should be a few years (e.g.,5 years) to remove the effect
of initial condition on the estimated results.}

\item{printAll}{A logical variable, if TRUE then the results (simulated isotope
concentration in streamflow) is printed for the entire simulation period,
including the warm up period. If FALSE then the results is only for the
simulatedDate.}
}
\value{
A data frame containing the date and simulated tracer concentrations.
}
\description{
Convolve the sine-wave function with the gamma distribution function
}
\details{
This function convolute the input sine wave (cP = AP*sin(2*pi*t - phiP)
+ kP) with the given gamma distribution (user given the shape and scale factor)
and return the simulated tracer concentrations in streamflow.
}
\examples{
# Get tracer data in streamflow of the Alp catchment (from the example dataset)
isotopeS_Alp <- subset(isotopeData, catchment == "Alp" & variable == "streamflow")

\dontrun{
simC <- convolSineNL(AP = 1.909, phiP = 2.016, kP = -11.05995, estAlpha = 0.802,
                     estBeta = 0.127, simulatedDate = isotopeS_Alp$date,
                     nWarmupYears = 5, printAll = FALSE)
}
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297
}
