% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlphaBeta.R
\name{findAlphaBeta}
\alias{findAlphaBeta}
\title{Find findAlphaBeta using bisection approach}
\usage{
findAlphaBeta(phiS_phiP = NULL, AS = NULL, AP = NULL, eps = 1e-06)
}
\arguments{
\item{phiS_phiP}{Phase shift of isotope concentration in streamflow (in radian). \cr
phiS_phiP MUST be > 0}

\item{AS}{The amplitude of the sine-wave function of isotope concentration in
precipitation. AS MUST be > 0}

\item{AP}{The amplitude of the sine-wave function of isotope concentration in
streamflow. AP MUST be > 0}

\item{eps}{The minimum search distance}
}
\value{
A list object containing the shape (alpha), scale (beta) factors,
and also mean transit time
}
\description{
Find alpha (shape) and beta (scale) factors of the gamma
distriubtion using the bisection approach. Solve equations (10) and (11) in
Kirchner (2016).
}
\examples{
result <- findAlphaBeta(phiS_phiP = pi,
                        AS = 1.0,
                        AP = 2.0,
                        eps = 1e-6)
# alpha (shape) factor
result$alpha
# beta (scale) factor
result$beta
# Mean transit time (years)
result$meanTT
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
