% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPhi.R
\name{findPhi}
\alias{findPhi}
\title{Find phiP within the range of 0 and 2pi}
\usage{
findPhi(aP = NULL, bP = NULL)
}
\arguments{
\item{aP}{A scalar or vector of the shape factor of the gamma distribution}

\item{bP}{A scalar or vector of the scale factor of the gamma distribution
phiS MUST be smaller than phiP}
}
\value{
A single phiP value within the raange of [0, 2*pi]
}
\description{
Find phiP within the range of 0 and 2pi
}
\details{
This function find root of the following system of equations for phiP
that within the range of [0, 2*pi] given values of aP and bP
aP <- - sqrt(aP^2 + bP^2) * sin(phiP) = -Ap*sin(phiP)
bP <- sqrt(aP^2 + bP^2) * cos(phiP) = Ap*cos(phiP)
}
\examples{
Ap <- 2.5
phiP <- 0.1
findPhi(aP = -Ap*sin(phiP), bP = Ap*cos(phiP))
# This function should return phiP = 0.1
phiP <- 2*pi - 0.1
findPhi(aP = -Ap*sin(phiP), bP = Ap*cos(phiP))
# This function should return phiP = 2*pi - 0.1
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
