% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTauFyw.R
\name{findtauyw}
\alias{findtauyw}
\title{Find findtauyw using bisection approach}
\usage{
findtauyw(shape = NULL, scale = NULL, Fyw = NULL)
}
\arguments{
\item{shape}{Shape factor (or alpha) of the gamma distribution}

\item{scale}{Scale factor (or beta) of the gamma distribution
phiS MUST be smaller than phiP}

\item{Fyw}{Youngwater fraction, equals to the amplitude ratio AS/AP where AS
and AP are the amplitudes of the fitted sine wave to isotope concentrations in
streamflow and in precipitation, respectively.}
}
\value{
Return the age threshold of youngwater fraction tauyw in years.
}
\description{
Find findtauyw using bisection approach
}
\details{
In this function, we solve equation 13 (Kirchner et al., 2013) with
given shape (alpha), scale (beta), and youngwater fraction (Fyw). The result is
the age threshold of youngwater fraction (tauyw) in years
}
\examples{
findtauyw(shape = 1.0, scale = 2.0, Fyw = 0.25)
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
