% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGamma.R
\name{fitGamma}
\alias{fitGamma}
\title{Fit the gamma distribution to observed instream isotope.}
\usage{
fitGamma(
  AP = NULL,
  phiP = NULL,
  kP = NULL,
  alphaRange = NULL,
  betaRange = NULL,
  simulatedDate = NULL,
  fittedData = NULL,
  weight = NULL,
  nIter = 5000,
  nBestIter = 30,
  nCores = 1,
  nWarmupYears = NULL,
  setSeed = NULL
)
}
\arguments{
\item{AP}{Amplitude of the sine wave \eqn{(c_P(t) = A_P \cdot sin(2 \pi t - \varphi_P) + k_P)} of
isotope concentration in precipitation. This can be a scalar or a vector}

\item{phiP}{Phase of the sine wave \eqn{(c_P(t) = A_P \cdot sin(2 \pi t - \varphi_P) + k_P)} of
isotope concentration in precipitation. This can be a scalar or a vector}

\item{kP}{The constant of the sine wave \eqn{(c_P(t) = A_P \cdot sin(2 \pi t - \varphi_P) + k_P)} of
isotope concentration in precipitation. This can be a scalar or a vector.
If AP, phiP, kP are vectors, they must have the same length}

\item{alphaRange}{The range of alpha (shape) parameter of the gamma distribution}

\item{betaRange}{The range of beta (scale) parameter of the gamma distribution}

\item{simulatedDate}{The date vector (yyyy-mm-dd format as.Date), indicating
the period of simulation}

\item{fittedData}{The fitted isotope concentration in streamflow for the simulatedDate.
This must have the same length as simulatedDate and contains no missing data.
The model will calculate the mean square error between the simulated and fitted data
This mean square error could be weighted if weight is given}

\item{weight}{The weight vectors (e.g., streamflow volume) at the simulatedDate.
Must be a vector of the same length with simulatedDate.}

\item{nIter}{Number of iterations. This function will generate a random nIter
parametersets (alpha, beta) within the user-defined range as above. These
parametersets will be combined with a random parameterset of AP[i], phiP[i],
and kP[i]. The mean square error (weighted if the weights are given) between
simulated and fittedData, the nBestIter iteration will be saved as output from
this function.}

\item{nBestIter}{Number of best parameters that will be save to output}

\item{nCores}{Number of cores for parallel simulation, e.g., if nIter = 10 and
nCores = 2, then there will be 2 parallel run on each cores with 5 iterations}

\item{nWarmupYears}{Number of warmup years.}

\item{setSeed}{An integer number used for set.seed() to have a reproducible
results.}
}
\value{
A list object containing table of simulated results, parameter
set generated by LHS, weighted mean square error (all are only for the
nBestIter best parameter sets).
}
\description{
Find parameter of the gamma distribution by convoluting the fitted
sine-wave to precipitation with the gamma distribution. Here, parameters of the
gamma distribution is generated using uniform Latin Hypercube sampling. The
goodness-of-fit is evaluated by the mean (weighted or unweighted using streamflow
volume) square error between observed instream isotope and simulated instream isotope.
}
\details{
This function perform the following convolution equation to find the simulated
isotope in stream (please see Eq. 1 from Kirchner (2016)):

\deqn{c_S(t) = \int_{0}^{\infty} h(\tau) \cdot c_P(t - \tau) d\tau }{}

where \eqn{c_P} in the following sine-wave form: \cr

 \deqn{c_P = A_P \cdot sin(2 \pi t - \varphi_P) + k_P}{}

and the gamma distribution is:

\deqn{h(\tau) =  \frac{\tau^{\alpha - 1}}{\beta^{\alpha} \Gamma (\alpha)} e^ {-t/\beta}}{}

In this convolution, parameters of the sine-wave (which is the fitted sine-wave
to isotope in precipitation) are given as input to this convolution. Here, a number
of random parameter sets for the gamma are generated using uniform Latin Hypercube
Sampling and these sampled parameter sets are randomly combined with parameter sets
of the sine-wave.
}
\examples{

#Get isotope data in streamflow of the Alp catchment (from the example dataset)
isotopeS_Alp <- subset(isotopeData, catchment == "Alp" & variable == "streamflow")

simC <- fitGamma(AP = 1.909, phiP = 2.016, kP = -11.05995, alphaRange = c(0.01,5),
                     betaRange = c(0.01,10), simulatedDate = isotopeS_Alp$date,
                     fittedData = isotopeS_Alp$delta_18O,
                     weight = isotopeS_Alp$water_flux_mm, nIter = 8,
                     nBestIter = 3, nCores = 1,
                     nWarmupYears = 5)
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems –
Part 1: Seasonal tracer cycles quantify young water fractions, but not mean
transit times, in spatially heterogeneous catchments. \emph{Hydrol. Earth Syst. Sci.},
20, 279–297.
}
