% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGamma.R
\name{fitGamma}
\alias{fitGamma}
\title{Fit gamma distribution to observed instream tracer concentratoin}
\usage{
fitGamma(
  AP = NULL,
  phiP = NULL,
  kP = NULL,
  alphaRange = NULL,
  betaRange = NULL,
  simulatedDate = NULL,
  fittedData = NULL,
  weight = NULL,
  nIter = 5000,
  nBestIter = 30,
  nCores = 1,
  nWarmupYears = NULL,
  setSeed = NULL
)
}
\arguments{
\item{AP}{Amplitude of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
isotope concentration in precipitation. This can be a scalar or a vector}

\item{phiP}{Phase of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
isotope concentration in precipitation. This can be a scalar or a vector}

\item{kP}{The constant of the sine wave (cP = AP*sin(2*pi*t - phiP) + kP) of
isotope concentration in precipitation. This can be a scalar or a vector.
If AP, phiP, kP are vectors, they must have the same length}

\item{alphaRange}{The range of alpha (shape) parameter of the gamma distribution}

\item{betaRange}{The range of beta (scale) parameter of the gamma distribution}

\item{simulatedDate}{The date vector (yyyy-mm-dd format as.Date), indicating
the period of simulation}

\item{fittedData}{The fitted isotope concentration in streamflow for the simulatedDate.
This must have the same length as simulatedDate and contains no missing data.
The model will calculate the mean square error between the simulated and fitted data
This mean square error could be weighted if weight is given}

\item{weight}{The weight vectors (e.g., streamflow volume) at the simulatedDate.
Must be a vector of the same length with simulatedDate.}

\item{nIter}{Number of iterations. This function will generate a random nIter
parametersets (alpha, beta) within the user-defined range as above. These
parametersets will be combined with a random parameterset of AP[i], phiP[i],
and kP[i]. The mean square error (weighted if the weights are given) between
simulated and fittedData, the nBestIter iteration will be saved as output from
this function.}

\item{nBestIter}{Number of best parameters that will be save to output}

\item{nCores}{Number of cores for parallel simulation, e.g., if nIter = 10 and
nCores = 2, then there will be 2 parallel run on each cores with 5 iterations}

\item{nWarmupYears}{Number of warmup years}

\item{setSeed}{An integer number used for set.seed() to have a reproducible
results.}
}
\value{
A list object containing a tibble objects of simulated results, parameter
set, weighted square error
}
\description{
Fit gamma distribution to observed instream tracer concentratoin
}
\details{
This function convolute the input sine wave (cP = AP*sin(2*pi*t - phiP)
+ kP) with the given gamma distribution (user given the shape and scale factor)
and return the simulated isotope concentrations in streamflow
}
\examples{

#Get isotope data in streamflow of the Alp catchment (from the example dataset)
isotopeS_Alp <- subset(isotopeData, catchment == "Alp" & variable == "streamflow")

simC <- fitGamma(AP = 1.909, phiP = 2.016, kP = -11.05995, alphaRange = c(0.01,5),
                     betaRange = c(0.01,10), simulatedDate = isotopeS_Alp$date,
                     fittedData = isotopeS_Alp$delta_18O,
                     weight = isotopeS_Alp$water_flux_mm, nIter = 8,
                     nBestIter = 3, nCores = 1,
                     nWarmupYears = 5)
}
\references{
Kirchner, J. W. (2016). Aggregation in environmental systems – Part 1: Seasonal
tracer cycles quantify young water fractions, but not mean transit times,
in spatially heterogeneous catchments, Hydrol. Earth Syst. Sci., 20, 279–297,
https://doi.org/10.5194/hess-20-279-2016.
}
