% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopeData}
\alias{isotopeData}
\title{Isotope (O18) in precipitation and streamflow in the Alp and Erlenbach
catchments in Switzerland}
\description{
Isotope (O18) in precipitation and streamflow in the Alp and Erlenbach
catchments in Switzerland
}
\references{
\url{https://www.envidat.ch/#/metadata/four-years-of-daily-stable-water-isotope-data}
}
\author{
Tam V. Nguyen \email{nguyen@ufz.de}
}
\keyword{Isotope}
\keyword{and}
\keyword{in}
\keyword{precipitation}
\keyword{streamflow}
